+++ lib/ui/onboarding/OnboardingLoginPhonePanel.dart
@@ -22,7 +22,7 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/ui/onboarding/OnboardingLoginPhoneVerifyPanel.dart';
 import 'package:illinois/ui/onboarding/OnboardingBackButton.dart';
 import 'package:illinois/service/Styles.dart';
-import 'package:illinois/ui/onboarding/onboarding2/Onboarding2Widgets.dart';
+import 'package:illinois/ui/onboarding2/Onboarding2Widgets.dart';
 import 'package:illinois/ui/widgets/ScalableWidgets.dart';
 
 class OnboardingLoginPhonePanel extends StatefulWidget with OnboardingPanel {