+++ lib/ui/onboarding2/Onboadring2RolesPanel.dart
@@ -21,7 +21,7 @@ import 'package:illinois/service/Auth2.dart';
 import 'package:illinois/service/Onboarding2.dart';
 import 'package:illinois/service/Localization.dart';
 import 'package:illinois/service/Analytics.dart';
-import 'package:illinois/ui/onboarding/onboarding2/Onboarding2PrivacyStatementPanel.dart';
+import 'package:illinois/ui/onboarding2/Onboarding2PrivacyStatementPanel.dart';
 import 'package:illinois/ui/widgets/RoleGridButton.dart';
 import 'package:illinois/service/Styles.dart';
 import 'package:illinois/ui/widgets/ScalableWidgets.dart';